PBCS-22: Show similar fields on Attributes and in editor. Show headings in bold
