+++ ui/widgets/widgets-dir/et-strapi-template-widget/src/integration/StrapiAPI.js
@@ -1,35 +1,8 @@
 import axios from "axios";
-import { KC_TOKEN_PREFIX, STRAPI_COLTYPE_URL, STRAPI_CONTYPE_URL } from "../constant/constant";
+import { KC_TOKEN_PREFIX } from "../constant/constant";
 import { addAuthorizationRequestConfig } from "./Integration";
 
 const strapiBaseUrl = `${process.env.REACT_APP_STRAPI_API_URL}`;
-
-/**
- * Get attribute fields of given content type from strapi
- * @param {*} contentType 
- * @returns 
- */
-export const getFields = async (contentType) => {
-    const url = `${strapiBaseUrl}/content-manager/collection-types/${contentType}`; //TODO: use this through proxy 
-    // const { data: { results } } = await axios.get(`${STRAPI_COLTYPE_URL}${contentType}`, addAuthorizationRequestConfig({}, KC_TOKEN_PREFIX));
-    const { data: { results } } = await axios.get(url, addAuthorizationRequestConfig({}, KC_TOKEN_PREFIX));
-
-    const content = {};
-    if (results && results.length) {
-        const fieldsArr = Object.keys(results[0]);
-
-        fieldsArr.map((el) => {
-            content[el + "}}"] = [
-                "getTextForLang(\"<LANG_CODE>\")",
-                "text",
-                "textMap(\"<LANG_CODE>\")"
-            ]
-        });
-    }
-    let contentObject = { 'content': content }
-    return contentObject;
-}
-
 
 /*********************
  * Strapi COLLECTION TYPE
@@ -54,7 +27,7 @@ export const getFields = async (contentType) => {
  * @returns 
  */
 export const getStrapiContentTypes = async () => {
-    const url = `${strapiBaseUrl}/content-manager/content-types`; //TODO: use this through proxy
+    const url = `${strapiBaseUrl}/content-manager/content-types`;
     // const data = await axios.get(STRAPI_CONTYPE_URL, addAuthorizationRequestConfig({}, KC_TOKEN_PREFIX));
     const data = await axios.get(url, addAuthorizationRequestConfig({}, KC_TOKEN_PREFIX));
     return data;
